+++ example/SubApp/Angular/extra-webpack.config.js
@@ -1,0 +1,9 @@
+const osAngularWebpack = require('@alicloud/console-os-ng-builder').default;
+
+module.exports = osAngularWebpack({}, {
+  id: 'os-exmaple-angular'
+});
+
+// module.exports = (config, ops) => {
+//   console.log(config)
+// }