+++ example/SubApp/Angular/src/app/product-list/product-list.component.ts
@@ -1,0 +1,21 @@
+import { Component, Injectable } from '@angular/core';
+import { products } from '../products';
+
+@Component({
+  selector: 'app-product-list',
+  templateUrl: './product-list.component.html',
+  styleUrls: ['./product-list.component.css']
+})
+@Injectable({
+  providedIn: 'root'
+})
+export class ProductListComponent {
+  products = products;
+}
+
+
+/*
+Copyright Google LLC. All Rights Reserved.
+Use of this source code is governed by an MIT-style license that
+can be found in the LICENSE file at http://angular.io/license
+*/