+++ packages/react-portal/src/mount.tsx
@@ -31,7 +31,7 @@ const unbindEvents = (emitter: EventEmitter) => {
 }
 
 const getProps = (props) => {
-  const appProps = { ...props };
+  const appProps = { ...props, ...(props.appProps || {}) };
 
   delete appProps.domElement;
   delete appProps.singleSpa;