+++ CHANGELOG.md
@@ -1,2 +1,8 @@
+# 0.1.2
+Fix: fix typo error in react-application: unmout -> unmount
+
+# 0.1.1
+Fix: fix typo error in react-application: unmout -> unmount
+
 # 0.1.0
 First Release. Introduction see [AlibabaCloud Console OS](https://aliyun.github.io/console-os)